requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
APScheduler==3.11.0
attrs==25.3.0
backports.ssl_match_hostname==3.7.0.1
certifi==2025.7.9
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
DataProperty==1.1.0
datasets==3.1.0
dill==0.3.8
docker==7.1.0
filelock==3.18.0
flwr==1.20.0
flwr-datasets==0.5.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.9.0
grpcio==1.73.1
grpcio-health-checking==1.62.3
hf-xet==1.1.5
huggingface-hub==0.33.4
humanreadable==0.4.1
hydra-core==1.3.2
idna==3.10
iniconfig==2.1.0
iterators==0.0.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mbstrdecoder==1.1.4
mdurl==0.1.2
mpmath==1.3.0
msgfy==0.2.1
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==25.0
pandas==2.3.1
path==17.1.1
pathspec==0.12.1
pathvalidate==3.3.1
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pyarrow==20.0.0
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pyparsing==3.2.3
pyroute2==0.9.4
pytest==8.4.1
python-dateutil==2.9.0.post0
python-iptables==1.2.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
ray==2.47.1
referencing==0.36.2
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
seaborn==0.13.2
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
SimpleSQLite==1.5.4
six==1.17.0
sqliteschema==2.0.1
subprocrunner==2.0.1
sympy==1.14.0
tabledata==1.3.4
tcconfig==0.30.1
tomli==2.2.1
tomli_w==1.2.0
topohub==1.4.1
torch==2.7.1
torchtext==0.6.0
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typepy==1.3.4
typer==0.12.5
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
voluptuous==0.15.2
xxhash==3.5.0
yarl==1.20.1
